pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voice-coder"
version = "0.3.0"
description = "项目感知型语音输入工具，支持动态热词学习"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "VoiceCoder Team" }
]
dependencies = [
    "vosk>=0.3.45",
    "pyaudio>=0.2.14",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "watchdog>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
voice-coder = "voice_coder.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
